Add tests for BandwidthMaintenanceSessionCleanup
